+++ src/entropy.rs
@@ -9,15 +9,12 @@
 
 use std::fmt;
 
-use bip39::rand::rngs::OsRng;
-use bip39::{Language, Mnemonic as Bip39Mnemonic};
-
 use crate::config::WALLET_KEYS_SEED_LEN;
-use crate::ffi::{maybe_deref, maybe_wrap};
+use crate::ffi::maybe_deref;
 use crate::io;
 
 #[cfg(not(feature = "uniffi"))]
-type Mnemonic = Bip39Mnemonic;
+type Mnemonic = bip39::Mnemonic;
 #[cfg(feature = "uniffi")]
 type Mnemonic = std::sync::Arc<crate::ffi::Mnemonic>;
 
@@ -122,91 +119,5 @@ impl fmt::Display for NodeEntropy {
 impl fmt::Debug for NodeEntropy {
 	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
 		write!(f, "NODE ENTROPY")
-	}
-}
-
-/// Generates a random [BIP 39] mnemonic with the specified word count.
-///
-/// If no word count is specified, defaults to 24 words (256-bit entropy).
-///
-/// The result may be used to initialize the [`NodeEntropy`], i.e., can be given to
-/// [`NodeEntropy::from_bip39_mnemonic`].
-///
-/// [BIP 39]: https://github.com/bitcoin/bips/blob/master/bip-0039.mediawiki
-/// [`Node`]: crate::Node
-pub fn generate_entropy_mnemonic(word_count: Option<WordCount>) -> Mnemonic {
-	let word_count = word_count.unwrap_or(WordCount::Words24).word_count();
-	let mnemonic = Bip39Mnemonic::generate_in_with(&mut OsRng, Language::English, word_count)
-		.expect("Failed to generate mnemonic");
-	maybe_wrap(mnemonic)
-}
-
-/// Supported BIP39 mnemonic word counts for entropy generation.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[cfg_attr(feature = "uniffi", derive(uniffi::Enum))]
-pub enum WordCount {
-	/// 12-word mnemonic (128-bit entropy)
-	Words12,
-	/// 15-word mnemonic (160-bit entropy)
-	Words15,
-	/// 18-word mnemonic (192-bit entropy)
-	Words18,
-	/// 21-word mnemonic (224-bit entropy)
-	Words21,
-	/// 24-word mnemonic (256-bit entropy)
-	Words24,
-}
-
-impl WordCount {
-	/// Returns the word count as a usize value.
-	pub fn word_count(&self) -> usize {
-		match self {
-			WordCount::Words12 => 12,
-			WordCount::Words15 => 15,
-			WordCount::Words18 => 18,
-			WordCount::Words21 => 21,
-			WordCount::Words24 => 24,
-		}
-	}
-}
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-
-	#[test]
-	fn mnemonic_to_entropy_to_mnemonic() {
-		// Test default (24 words)
-		let mnemonic = generate_entropy_mnemonic(None);
-		let mnemonic_inner = maybe_deref(&mnemonic);
-		let entropy = mnemonic_inner.to_entropy();
-		assert_eq!(mnemonic_inner, &Bip39Mnemonic::from_entropy(&entropy).unwrap());
-		assert_eq!(mnemonic_inner.word_count(), 24);
-
-		// Test with different word counts
-		let word_counts = [
-			WordCount::Words12,
-			WordCount::Words15,
-			WordCount::Words18,
-			WordCount::Words21,
-			WordCount::Words24,
-		];
-
-		for word_count in word_counts {
-			let mnemonic = generate_entropy_mnemonic(Some(word_count));
-			let mnemonic_inner = maybe_deref(&mnemonic);
-			let entropy = mnemonic_inner.to_entropy();
-			assert_eq!(mnemonic_inner, &Bip39Mnemonic::from_entropy(&entropy).unwrap());
-
-			// Verify expected word count
-			let expected_words = match word_count {
-				WordCount::Words12 => 12,
-				WordCount::Words15 => 15,
-				WordCount::Words18 => 18,
-				WordCount::Words21 => 21,
-				WordCount::Words24 => 24,
-			};
-			assert_eq!(mnemonic_inner.word_count(), expected_words);
-		}
 	}
 }